requirements.txt
Django==2.2.6
djangorestframework==3.10.3
Markdown==3.1.1
django-filter==2.2.0
django-allauth==0.40.0
Pillow==6.2.0
psycopg2==2.8.3
django-rest-auth==0.9.5